+++ android/src/main/java/ly/count/dart/countly_flutter/CountlyFlutterPlugin.java
@@ -60,7 +60,7 @@ import ly.count.android.sdk.messaging.CountlyPush;
  */
 public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, ActivityAware, DefaultLifecycleObserver {
     private static final String TAG = "CountlyFlutterPlugin";
-    private final String COUNTLY_FLUTTER_SDK_VERSION_STRING = "24.7.1";
+    private final String COUNTLY_FLUTTER_SDK_VERSION_STRING = "24.11.0";
     private final String COUNTLY_FLUTTER_SDK_NAME = "dart-flutterb-android";
     private final String COUNTLY_FLUTTER_SDK_NAME_NO_PUSH = "dart-flutterbnp-android";
 
@@ -433,7 +433,25 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
                     }
                 }
                 CountlyPush.init(activity.getApplication(), pushTokenType);
+<<<<<<< HEAD
                 result.success(" askForNotificationPermission!");
+=======
+                FirebaseApp.initializeApp(context);
+                FirebaseMessaging.getInstance().getToken().addOnCompleteListener(new OnCompleteListener<String>() {
+                    @Override
+                    public void onComplete(@NonNull Task<String> task) {
+                        if (!task.isSuccessful()) {
+                            log("[askForNotificationPermission], Fetching FCM registration token failed", task.getException(), LogLevel.WARNING);
+                            return;
+                        }
+
+                        String token = task.getResult();
+                        log("FCM Token: " + token, LogLevel.INFO);
+                        CountlyPush.onTokenRefresh(token);
+                    }
+                });
+                result.success("askForNotificationPermission!");
+>>>>>>> 5c220d7e2d148de3de79d300a48a258fa44138d9
             } else if ("pushTokenType".equals(call.method)) {
                 String tokenType = args.getString(0);
                 if ("2".equals(tokenType)) {
@@ -496,12 +514,14 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
                 String key = args.getString(0);
                 int count = Integer.parseInt(args.getString(1));
                 float sum = Float.parseFloat(args.getString(2));
-                HashMap<String, Object> segmentation = new HashMap<>();
+                Map<String, Object> segmentation;
                 if (args.length() > 3) {
-                    for (int i = 3, il = args.length(); i < il; i += 2) {
-                        segmentation.put(args.getString(i), args.getString(i + 1));
-                    }
+                    segmentation = toMap(args.getJSONObject(3));
+                } else {
+                    segmentation = null;
                 }
+                // Map<String, Object> segmentation = toMap(args.getJSONObject(4));
+                
                 Countly.sharedInstance().events().endEvent(key, segmentation, count, sum);
                 result.success("endEvent for: " + key);
             } else if ("recordEvent".equals(call.method)) {
@@ -1225,7 +1245,7 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
                 this.config.setRecordAppStartTime(true);
                 result.success("enableApm: success");
             } else if ("throwNativeException".equals(call.method)) {
-                throw new IllegalStateException("Native Exception Crashhh!");
+                triggerOverflow();
             } else if ("recordIndirectAttribution".equals(call.method)) {
                 JSONObject attributionValues = args.getJSONObject(0);
                 if (attributionValues != null && attributionValues.length() > 0) {
@@ -1319,6 +1339,12 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
             } else if ("halt".equals(call.method)) {
                 Countly.sharedInstance().halt();
                 result.success("halt: success");
+            } else if ("enterContentZone".equals(call.method)) {
+                Countly.sharedInstance().contents().enterContentZone();
+                result.success(null);
+            } else if ("exitContentZone".equals(call.method)) {
+                Countly.sharedInstance().contents().exitContentZone();
+                result.success(null);
             }
             //------------------End------------------------------------
 
@@ -1329,6 +1355,14 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
         } catch (JSONException jsonException) {
             result.success(jsonException.toString());
         }
+    }
+
+    void triggerOverflow() {
+        triggerOverflow1();
+    }
+
+    void triggerOverflow1() {
+        triggerOverflow();
     }
 
     CountlyFeedbackWidget getFeedbackWidget(String widgetId) {
@@ -1497,6 +1531,11 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
         if (_config.has("httpPostForced")) {
             this.config.setHttpPostForced(_config.getBoolean("httpPostForced"));
         }
+        if (_config.has("customNetworkRequestHeaders")) {
+            JSONObject customHeaderValues = _config.getJSONObject("customNetworkRequestHeaders");
+            Map<String, String> customHeaderValuesMap = toMapString(customHeaderValues);
+            this.config.addCustomNetworkRequestHeaders(customHeaderValuesMap);
+        }
         if (_config.has("shouldRequireConsent")) {
             this.config.setRequiresConsent(_config.getBoolean("shouldRequireConsent"));
         }
@@ -1654,6 +1693,13 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
 
         if (_config.has("autoEnrollABOnDownload") && _config.getBoolean("autoEnrollABOnDownload")) {
              this.config.enrollABOnRCDownload();
+        }
+
+        if (_config.has("visibilityTracking")) {
+            this.config.experimental.enableVisibilityTracking();
+        }
+        if (_config.has("previousNameRecording")) {
+            this.config.experimental.enablePreviousNameRecording();
         }
     }
 }