+++ example/integration_test/session_tests/automatic_sessions/206_CR_CNG_A_id_change_test.dart
@@ -63,10 +63,10 @@ void main() {
         // example:
         // consent: [{"sessions":true,"crashes":true,"users":true,"push":true,"feedback":true,"scrolls":true,"remote-config":true,"attribution":true,"clicks":true,"location":true,"star-rating":true,"events":true,"views":true,"apm":true}]
         Map<String, dynamic> consentInRequest = jsonDecode(queryParams['consent']![0]);
-        for (var key in ['push', 'feedback', 'crashes', 'attribution', 'users', 'events', 'remote-config', 'sessions', 'location', 'views', 'apm']) {
+        for (var key in ['push', 'feedback', 'crashes', 'attribution', 'users', 'events', 'remote-config', 'sessions', 'location', 'views', 'apm', 'content']) {
           expect(consentInRequest[key], false);
         }
-        expect(consentInRequest.length, Platform.isAndroid ? 14 : 11);
+        expect(consentInRequest.length, Platform.isAndroid ? 15 : 12);
       } else if (i == 2 || i == 3 || i == 4) {
         expect(queryParams['old_device_id']?[0].isNotEmpty, true);
         expect(queryParams['device_id']?[0], 'newID');