+++ example/integration_test/utils.dart
@@ -25,11 +25,51 @@ Future<List<String>> getEventQueue() async {
   return eq.cast<String>();
 }
 
+/// Get event queue from native side (list of json objects)
+Future<List<dynamic>> getEventsFromEventQueueAndRequestList() async {
+  List<String> requestList = await getRequestQueue();
+  List<String> eventList = await getEventQueue();
+  // Filter requestList to include only requests containing 'events_'
+  List<String> filteredRequestList = requestList.where((element) => element.contains('events=')).toList();
+
+  // Some logs for debugging
+  print('Filtered Request List: $filteredRequestList');
+
+  // Process each filtered request
+  List<dynamic> events = [];
+
+  for (var element in filteredRequestList) {
+    Map<String, List<String>> queryParams = Uri
+        .parse('?' + element)
+        .queryParametersAll;
+
+    if (queryParams.containsKey('events')) {
+      var eventsEncoded = queryParams['events']!.first; // Get the first item as it is a List<String>
+      var eventsRaw = json.decode(Uri.decodeComponent(eventsEncoded));
+
+      if (eventsRaw is List) {
+        for (var event in eventsRaw) {
+          events.add(event);
+        }
+      }
+    }
+  }
+
+  print('Filtered Events Request List: ${events.length}');
+
+  for (var index = 0; index < eventList.length; index++) {
+    Map<String, dynamic> event = json.decode(eventList[index]);
+    events.add(event);
+  }
+
+  return events;
+}
+
 /// Verify the common request queue parameters
 void testCommonRequestParams(Map<String, List<String>> requestObject) {
   expect(requestObject['app_key']?[0], APP_KEY);
   expect(requestObject['sdk_name']?[0], "dart-flutterbnp-${Platform.isIOS ? "ios" : "android"}");
-  expect(requestObject['sdk_version']?[0], '24.7.1');
+  expect(requestObject['sdk_version']?[0], '24.11.0');
   expect(requestObject['av']?[0], Platform.isIOS ? '0.0.1' : '1.0.0');
   assert(requestObject['timestamp']?[0] != null);
 
@@ -172,6 +212,24 @@ Future<Map<String, dynamic>> getApmParamsFromRequest(String request) async {
   return apmParams;
 }
 
+/// Go to background
+void goBackground() {
+  FlutterForegroundTask.minimizeApp();
+  if (Platform.isIOS) {
+    printMessageMultipleTimes('will now go to background, get ready to go foreground manually', 3);
+  }
+  sleep(Duration(seconds: 3));
+}
+
+/// Go to foreground
+void goForeground() {
+  FlutterForegroundTask.launchApp();
+  if (Platform.isIOS) {
+    printMessageMultipleTimes('waiting for 3 seconds, now go to foreground', 3);
+  }
+  sleep(Duration(seconds: 3));
+}
+
 /// Go to background and foreground
 void goBackgroundAndForeground() {
   FlutterForegroundTask.minimizeApp();
@@ -194,6 +252,43 @@ void printMessageMultipleTimes(String message, int times) {
     print(message);
   }
 }
+
+Future<void> createMixViewsAndEvents(bool inForeground) async {
+  // Set global view segmentation
+  Map<String, Object> viewSegmentation = {
+    'isForeground': inForeground
+  };
+  await Countly.instance.views.setGlobalViewSegmentation(viewSegmentation);
+
+  // Create events with segmentation
+  var event1 = {'key': 'E1${inForeground ? '_FG' : '_BG'}'};
+  await Countly.recordEvent(event1);
+
+  await Countly.instance.views.startView("V1${inForeground ? '_FG' : '_BG'}");
+
+  var event2 = {'key': 'E2${inForeground ? '_FG' : '_BG'}', 'count': 3, 'sum': '2.99'};
+  await Countly.recordEvent(event2);
+
+  await Countly.instance.views.startAutoStoppedView("V2${inForeground ? '_FG' : '_BG'}");
+
+  var event3 = {'key': 'E3${inForeground ? '_FG' : '_BG'}'};
+  await Countly.recordEvent(event3);
+
+  await Countly.instance.views.startAutoStoppedView('V3${inForeground ? '_FG' : '_BG'}');
+
+  var event4 = {'key': 'E4${inForeground ? '_FG' : '_BG'}'};
+  await Countly.recordEvent(event4);
+
+  await Countly.instance.views.startView("V4${inForeground ? '_FG' : '_BG'}");
+
+  var event5 = {
+    'key': 'E5${inForeground ? '_FG' : '_BG'}',
+    'count': 5,
+    'sum': '5.99'
+  };
+  await Countly.recordEvent(event5);
+}
+
 
 /// Creates truncable events (which covers all possible truncable situations)
 /// - Events with segmentation